worker: Move FakeProtocolCommand class to separate file
